proxy-server,refactor: moving functions to helpers
- Created lib/helpers.js
- Moved printTime and padLeft to helpers
- exposed printTime
- Imported printTime from helpers to server.js

CLoses isue #70
